Put loading indicator in own module
